+++ spec/app/models/scheduler.rb
@@ -1,0 +1,7 @@
+class Scheduler
+  include Mongoid::Document
+
+  def strategy
+    Strategy.new
+  end
+end